net change preview of selection, tooltips
