+++ DEHPCommon/UserInterfaces/ViewModels/Interfaces/IObjectBrowserViewModel.cs
@@ -35,72 +35,7 @@ namespace DEHPCommon.UserInterfaces.ViewModels.Interfaces
     /// <summary>
     /// Interface definition for the <see cref="ObjectBrowserViewModel"/>
     /// </summary>
-    public interface IObjectBrowserViewModel
+    public interface IObjectBrowserViewModel : IObjectBrowserBaseViewModel
     {
-        /// <summary>
-        /// Gets or sets a value indicating whether the browser is busy
-        /// </summary>
-        bool? IsBusy { get; set; }
-
-        /// <summary>
-        /// Gets or sets the selected thing
-        /// </summary>
-        object SelectedThing { get; set; }
-
-        /// <summary>
-        /// Gets or sets the selected things collection
-        /// </summary>
-        ReactiveList<object> SelectedThings { get; set; }
-
-        /// <summary>
-        /// Gets the collection of <see cref="IRowViewModelBase{T}"/> to be displayed in the tree
-        /// </summary>
-        ReactiveList<BrowserViewModelBase> Things { get; }
-
-        /// <summary>
-        /// Gets the Context Menu for the implementing view model
-        /// </summary>
-        ReactiveList<ContextMenuItemViewModel> ContextMenu { get; }
-
-        /// <summary>
-        /// Gets the command that allows to map the selected things
-        /// </summary>
-        ReactiveCommand<object> MapCommand { get; set; }
-
-        /// <summary>
-        /// Gets or sets the <see cref="IObservable{T}"/> of <see cref="bool"/> that is bound to the <see cref="MapCommand"/> <see cref="ReactiveCommand{T}.CanExecute"/> property
-        /// </summary>
-        /// <remarks>This observable is intended to be Merged with another observable</remarks>
-        IObservable<bool> CanMap { get; set; }
-
-        /// <summary>
-        /// Gets the Caption of the control
-        /// </summary>
-        string Caption { get; }
-
-        /// <summary>
-        /// Gets the tooltip of the control
-        /// </summary>
-        string ToolTip { get; }
-
-        /// <summary>
-        /// Reloads the the trees elements
-        /// </summary>
-        void Reload();
-
-        /// <summary>
-        /// Adds to the <see cref="ObjectBrowserViewModel.Things"/> collection the specified by <see cref="IObjectBrowserTreeSelectorService"/> trees
-        /// </summary>
-        void BuildTrees();
-
-        /// <summary>
-        /// Populate the context menu for the implementing view model
-        /// </summary>
-        void PopulateContextMenu();
-
-        /// <summary>
-        /// Dispose of this <see cref="ObjectBrowserViewModel"/>
-        /// </summary>
-        void Dispose();
     }
 }
