+++ DEHPCommon/UserInterfaces/Views/TransferControl.xaml
@@ -23,7 +23,8 @@
             <ProgressBar Background="AliceBlue" Height="40" Foreground="DarkGray" Maximum="100" Minimum="0" IsIndeterminate="{Binding IsIndeterminate}" Value="{Binding Progress}" />
             <TextBlock HorizontalAlignment="Center" Visibility="{Binding IsIndeterminate, Converter={StaticResource BooleanToVisibilityConverter}}" VerticalAlignment="Center" FontSize="18" Text="{Binding Progress, StringFormat={}{0}%}" />
         </Grid>
-        <Button Grid.Row="1" Grid.Column="0" Height="40" Margin="10,5,5,10" Content="Cancel" Command="{Binding CancelCommand}"/>
-        <Button Grid.Row="1" Grid.Column="1" Height="40" Margin="5,5,10,10" Content="Transfer" Command="{Binding TransferCommand}"/>
+        <Button ToolTip="Attempt to cancel the transfers" Grid.Row="1" Grid.Column="0" Height="40" Margin="10,5,5,10" Content="Cancel" Command="{Binding CancelCommand}"/>
+        <Button ToolTip="Transfer all mapped things from dst to hub &#x0a;and mapped things from hub to dst &#x0a; Mapping will also be persisted"
+                Grid.Row="1" Grid.Column="1" Height="40" Margin="5,5,10,10" Content="Transfer" Command="{Binding TransferCommand}"/>
     </Grid>
 </UserControl>
